refactor: exporting all the backend env from a single file and reusing them
