+++ backend/config/utils.js
@@ -1,0 +1,8 @@
+import dotenv from 'dotenv';
+dotenv.config();
+
+const PORT = process.env.PORT;
+const MONGODB_URI = process.env.MONGODB_URI;
+const REDIS_URL = process.env.REDIS_URL;
+
+export { MONGODB_URI, PORT, REDIS_URL };
